pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-detection"
version = "1.0.0"
description = "End-to-End MLOps Pipeline for Fraud Detection"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "mlflow>=2.10",
    "prefect>=3.0",
    "pandera>=0.18",
    "imbalanced-learn>=0.12",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pydantic>=2.0",
    "evidently>=0.4",
    "prometheus-fastapi-instrumentator>=6.0",
    "prometheus-client>=0.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "httpx>=0.27",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
